Split up networking resources into shared and non-shared directories. Shared means that they are shared with the canal networking setup.
First I tried to use symbolic links to share files between canal and the
rest but this didn't work out:

Terraform 0.11.0 doesn't support symbolic links in the source directory of
the template_dir resource.  This is because changes in this directory are
detected by building a tar archive on the fly and calculating a hash over
it.  The tar archive building doesn't support symlinks currently.  This
could be fixed but would require changes in Terraform obviously.