Revert "Switch kube-proxy from iptables to IPVS"
This reverts commit c65c42716dbab21d4f68e05dd2aa80801aea22d1.

ipvs mode unfortunately doesn't work in one corner case: accessing
an in-cluster service via its external ip address when the pod
accessing the service and the pod delivering the service are on
different nodes and the externalTrafficPolicy is Local.

See https://github.com/kubernetes/kubernetes/issues/75262 for
a lengthy discussion about the topic.